fix(build): move upsertSubmissionByDraft to .server to keep postgres out of client bundle
public-submissions.ts is imported by client code (autosave hook, public form
page). The createServerFn body is stripped on the client, but the new top-level
upsertSubmissionByDraft export was not — it referenced @/db and dragged the
postgres driver into the client build, failing Vercel with
'"performance" is not exported by __vite-browser-external'. Extract it into
public-submissions-upsert.server.ts so the .server suffix excludes it.
